requirements.txt
matplotlib~=3.5.2
pandas~=1.4.2
Pillow==9.2.0
scikit_learn==1.1.2
streamlit==1.12.0
requests~=2.26.0
sklearn~=0.0
scikit-learn~=1.1.2